requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    ./dev-cli.py update
#
async-timeout==4.0.3 \
    --hash=sha256:4640d96be84d82d02ed59ea2b7105a0f7b33abe8703703cd0ab0bf87c427522f \
    --hash=sha256:7405140ff1230c310e51dc27b3145b9092d659ce68ff733fb0cefe3ee42be028
    # via cli-base-utilities
bx-py-utils==90 \
    --hash=sha256:2c15b3b7d8a8e4f0bc6815517fb732fb303ce6878c3e45bbd60633b2811bf0ea \
    --hash=sha256:bb8aa9bd59b6c733cc8fe65340a98bbff026bbc5aa3c3c01324f27d84b8c9302
    # via
    #   cli-base-utilities
    #   pysmartmeter (pyproject.toml)
cli-base-utilities==0.7.0 \
    --hash=sha256:3d13b861d4c96b597b1dc3c6fdbe428dc72c444ab61008cc9555292d7b3ec8b4 \
    --hash=sha256:fa39bd83900749ece5866ea996d270880081a5155497f74db46995c28b1a9425
    # via
    #   ha-services
    #   pysmartmeter (pyproject.toml)
click==8.1.7 \
    --hash=sha256:ae74fb96c20a0277a1d615f1e4d73c8414f5a98db8b799a7931d1582f3390c28 \
    --hash=sha256:ca9853ad459e787e2192211578cc907e7594e294c7ccc834310722b41b9ca6de
    # via
    #   cli-base-utilities
    #   ha-services
    #   pysmartmeter (pyproject.toml)
    #   rich-click
ha-services==0.4.0 \
    --hash=sha256:11085e9bccfedc0a561723eb084ab903c1301fe464d192df154c46a50726b5be \
    --hash=sha256:e3179926be454c6c3375e19bd545dbbb3f87876e3b95840fe8a469644eeb6791
    # via pysmartmeter (pyproject.toml)
markdown-it-py==3.0.0 \
    --hash=sha256:355216845c60bd96232cd8d8c40e8f9765cc86f46880e43a8fd22dc1a1a8cab1 \
    --hash=sha256:e3f60a94fa066dc52ec76661e37c851cb232d92f9886b15cb560aaada2df8feb
    # via rich
mdurl==0.1.2 \
    --hash=sha256:84008a41e51615a49fc9966191ff91509e3c40b939176e643fd50a5c2196b8f8 \
    --hash=sha256:bb413d29f5eea38f31dd4754dd7377d4465116fb207585f97bf925588687c1ba
    # via markdown-it-py
msgspec==0.18.5 \
    --hash=sha256:0017f6af35a3959002df4c82af60c1df2160701529dd89b17df971fde5945257 \
    --hash=sha256:0027fba5362a3cb1bdd5503709aa2dbffad22dffd50f415086ed5f74f229ead9 \
    --hash=sha256:01ed3472a0508f88a25a9d3bccafb840110f0fc5eb493b4baa43646e4e7c75c2 \
    --hash=sha256:024f880df7d2f8cfdb9f9904efa0f386d3692457159bd58f850c20f11c07d16f \
    --hash=sha256:13da9df61745b7757070dae6e3476ab4e13bb9dd3e3d11b050dfcae540058bd1 \
    --hash=sha256:261cc6e3a687e6f31b80056ab12f6adff3255f9b68b86d92b0b497f8b289c84c \
    --hash=sha256:3f387cabddf2dc26d6fa7f1a8158deefc8db9e0626eacebbe4875f421c66d574 \
    --hash=sha256:482bdf77f3892dd603061b2b21ac6a4492bb797a552c92e833a41fe157162257 \
    --hash=sha256:50479d88f3c4e9c73b55fbe84dc14b1cee8cec753e9170bbeafe3f9837e9f7af \
    --hash=sha256:6431305c645fb2a88a6da1fcec53dbaac61697f1219000b9589f9286532aabc0 \
    --hash=sha256:773a38ead7832d171d1b9406bf42448a218245584af36e42c31f26d9f48a493a \
    --hash=sha256:8e545651531f2d01b983d0ac0c7f3b6d99674267ff261b5f344f5016160b5608 \
    --hash=sha256:8f7c0460aefdc8f01ea35f26e38c62b574bbf0b138ade860f557bbf9e9dac50c \
    --hash=sha256:9420750f19c311e490db3edff9d153621c4989c582cf1be40c307c86d6cc2c1e \
    --hash=sha256:9a7b682cca3ba251a19cc769d38615ddd9551e086858decd950c156c2e79ecc1 \
    --hash=sha256:a0ec78bd93684db61dfccf7a421b2e1a525b1a0546b4d8c4e339151be57d58a6 \
    --hash=sha256:b3d206af4280172948d014d20b2cea7939784a99ea9a7ac943ce71100dbe8f98 \
    --hash=sha256:b491b2549d22e11d7cfe34a231f9bd006cb6b71adefa070a070075d2f601e75c \
    --hash=sha256:b547c7ad9786a79b0090a811d95d2d04063625a66fd96ed767cdfbabd8087c67 \
    --hash=sha256:b6af133ba491a09ef8dcbc2d9904bcec220247e2067bb75d5d6daa12e0739d6c \
    --hash=sha256:b7b49cba0577edc8ac166263b5fec3619fe5a267805cfc041bccaf8a0c58ef05 \
    --hash=sha256:bd8a64da668b4eeef4b21dcecc640ed6950db661e2ea42ae52bbac5a2dbffb3a \
    --hash=sha256:be2440fa5699e1b3062d17fdfd8c6a459d72bb4edbce403353af6f39c8c5a6fa \
    --hash=sha256:c13e0a510bbd00cb29d193fceff55d1e17a99c9f97284cdbe61c15496c2f7803 \
    --hash=sha256:c30fadc1a1118097920dd868e42469fed32c7078ca2feff2fc19e7c017065322 \
    --hash=sha256:c79e7115f0143688c5d866359e7b6b76dd1581a81c9aeac7805a9d6320e9f2ca \
    --hash=sha256:cf885edac512e464c70a5f4f93b6f778c83ea4b91d646b6d72f6f5ac950f268e \
    --hash=sha256:d318593e0ddc11b600552a470ec27baeb0b86a8e37903ac5ce7472ba0d6f7bf8 \
    --hash=sha256:d5999eb65646b131f439ebb07c22446e8976b7fd8a312dca09ce6fa2c21162bb \
    --hash=sha256:e4c2fc93a98afefd1a78e957ca63363a8e5fd1b58bf70a8d66413c8f2a4723a2 \
    --hash=sha256:eccba21248f90f332335b109e89685e79940367974812cd13975313f480f3dd8 \
    --hash=sha256:ee1f9414523d9a53744d21a6a2b6a636d9008be016963148a2646b38132e11dd \
    --hash=sha256:f290bfe7e21e8069890d101d8a060500b22a3aeb7860274644c4ec9240ddbedc \
    --hash=sha256:f4eeb22921ca6cdfbf17ca874eccbe23eb010c89ffb3017b628940c37d53ce4a \
    --hash=sha256:f55c4610cb0514aef8b35bfd0682f4cc2d7efd5e9b58acf30abd90b2a2376b5d \
    --hash=sha256:fae28faef5fd61847930d8e86fd83c18f991a338efd8fbf69c1d35d42c652f41
    # via ha-services
packaging==23.2 \
    --hash=sha256:048fb0e9405036518eaaf48a55953c750c11e1a1b68e0dd1a9d62ed0c092cfc5 \
    --hash=sha256:8c491190033a9af7e1d931d0b5dacc2ef47509b34dd0de67ed209b5203fc88c7
    # via cli-base-utilities
paho-mqtt==1.6.1 \
    --hash=sha256:2a8291c81623aec00372b5a85558a372c747cbca8e9934dfe218638b8eefc26f
    # via
    #   ha-services
    #   pysmartmeter (pyproject.toml)
pygments==2.17.2 \
    --hash=sha256:b27c2826c47d0f3219f29554824c30c5e8945175d888647acd804ddd04af846c \
    --hash=sha256:da46cec9fd2de5be3a8a784f434e4c4ab670b4ff54d605c4c2717e9d49c4c367
    # via rich
pyserial==3.5 \
    --hash=sha256:3c77e014170dfffbd816e6ffc205e9842efb10be9f58ec16d3e8675b4925cddb \
    --hash=sha256:c4451db6ba391ca6ca299fb3ec7bae67a5c55dde170964c7a14ceefec02f2cf0
    # via pysmartmeter (pyproject.toml)
rich==13.7.0 \
    --hash=sha256:5cb5123b5cf9ee70584244246816e9114227e0b98ad9176eede6ad54bf5403fa \
    --hash=sha256:6da14c108c4866ee9520bbffa71f6fe3962e193b7da68720583850cd4548e235
    # via
    #   cli-base-utilities
    #   ha-services
    #   pysmartmeter (pyproject.toml)
    #   rich-click
rich-click==1.7.2 \
    --hash=sha256:22f93439a3d65f4a04e07cd584f4d01d132d96899766af92ed287618156abbe2 \
    --hash=sha256:a42bcdcb8696c4ca7a3b1a39e1aba3d2cb64ad00690b4c022fdcb2cbccebc3fc
    # via
    #   cli-base-utilities
    #   ha-services
    #   pysmartmeter (pyproject.toml)
tomli==2.0.1 \
    --hash=sha256:939de3e7a6161af0c887ef91b7d41a53e7c5a1ca976325f429cb46ea9bc30ecc \
    --hash=sha256:de526c12914f0c550d15924c62d72abc48d6fe7364aa87328337a31007fe8a4f
    # via cli-base-utilities
tomlkit==0.12.3 \
    --hash=sha256:75baf5012d06501f07bee5bf8e801b9f343e7aac5a92581f20f80ce632e6b5a4 \
    --hash=sha256:b0a645a9156dc7cb5d3a1f0d4bab66db287fcb8e0430bdd4664a095ea16414ba
    # via
    #   cli-base-utilities
    #   ha-services
typing-extensions==4.9.0 \
    --hash=sha256:23478f88c37f27d76ac8aee6c905017a143b0b1b886c3c9f66bc2fd94f9f5783 \
    --hash=sha256:af72aea155e91adfc61c3ae9e0e342dbc0cba726d6cba4b6c72c1f34e47291cd
    # via rich-click
